Folder structure (#24)
* Fix typo import

* Folder structure
