+++ lib/constants/layout.dart
@@ -1,0 +1,1 @@
+const kBorderRadius = 8.0;
