+++ lib/screens/home_screen.dart
@@ -10,12 +10,12 @@ import '../screens/add_task.dart';
 import '../screens/settings_screen.dart';
 
 // Widgets Imports
-import '../widgets/stats_chart.dart';
-import '../widgets/your_stats.dart';
-import '../widgets/carousel_card.dart';
-import '../widgets/taskslist.dart';
-import '../widgets/sleep_cycle_column.dart';
-import '../widgets/initialiser.dart';
+import '../widgets/home/stats/stats_chart.dart';
+import '../widgets/home/stats/your_stats.dart';
+import '../widgets/home/carousel_card.dart';
+import '../widgets/home/taskslist.dart';
+import '../widgets/home/sleep_cycle_column.dart';
+import '../widgets/home/initialiser.dart';
 
 class HomeScreen extends StatefulWidget {
   @override
